+++ spalipy/utils.py
@@ -1,4 +1,7 @@
 import numpy as np
+import mmap
+import tempfile
+from time import time
 
 
 def _c_array_prep(array):
@@ -11,3 +14,37 @@ def _c_array_prep(array):
         return array
     else:
         return array.newbyteorder("=").byteswap(inplace=False)
+
+
+def _memmap_tryfree(obj):
+    """Attempt to release memory usage from a np.memmap object. Return True on success, else False."""
+    if isinstance(obj, np.memmap):
+        try:
+            obj._mmap.madvise(mmap.MADV_DONTNEED)
+            return True
+        except:
+            pass
+    return False
+
+
+def _memmap_create_temp(ndarray_to_save):
+    """Create and return temporary file np.memmap object using defaults as per tempfile.
+    Temp file will be unlinked on exit but will persist for use until memmap is garbage collected."""
+    if not isinstance(ndarray_to_save, np.ndarray):
+        raise ValueError("ndarray_to_save must be np.ndarray object.")
+    tmp_file = tempfile.NamedTemporaryFile(
+        mode='w+b',
+        prefix=rf'spalipy_{int(time())}_',
+        suffix=r'.dat',
+        delete=True,
+    )
+    tmp_file.write(ndarray_to_save.tobytes(order='C'))
+    tmp_file.seek(0)
+    tmp_memmap = np.memmap(
+        tmp_file,
+        mode='r',
+        dtype=ndarray_to_save.dtype,
+        shape=ndarray_to_save.shape,
+        order='C',
+    )
+    return tmp_memmap